Moved to common circuit
